Add service.go
